Introduced checkout()
